% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_ecdf.R
\name{plot.ecdf_comparison}
\alias{plot.ecdf_comparison}
\title{Plot the gene probability summaries that the observed expression data belongs
to the training data distributions}
\usage{
\method{plot}{ecdf_comparison}(x, summary.func = mean)
}
\arguments{
\item{x}{ecdf_comparison object}

\item{summary.func}{Function to summarize across subjects for each gene}
}
\description{
This function summarizes the probabilities across subjects for each gene and
plots the summary probabilities to a histogram. The red reference line at
50\% refers to a negligible difference between the observed and training data.
}
\author{
Lorin Towle-Miller
}
