% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsGSVA.R
\name{rsGSVA}
\alias{rsGSVA}
\title{Run GSVA algorithm while providing separate distribution for underlying kcdf}
\usage{
rsGSVA(X, gs, train_expr = NULL, min.sz = 1, max.sz = 500, verbose = TRUE)
}
\arguments{
\item{X}{matrix of expression data. Rows are genes and columns are subjects}

\item{gs}{list of gene sets}

\item{train_expr}{matrix of reference expression data. Rows are subjects and columns are genes}

\item{min.sz}{min.sz from \code{GSVA::filterGeneSets}}

\item{max.sz}{max.sz from \code{GSVA::filterGeneSets}}

\item{verbose}{TRUE/FALSE whether log messages should be shown}
}
\description{
This function estimates GSVA enrichment scores (Hänzelmann et al, 2013) with given
test and target distribution samples. Expects desired test matrix, gene set, target matrix and min
and max sizing for filtering gene sets.
}
\references{
Hänzelmann S, Castelo R, Guinney J. GSVA: gene set variation analysis for microarray and RNA-seq data. BMC bioinformatics. 2013 Dec;14:1-5.
}
\author{
William Jordan
}
